requirements.txt
# Python packages requirements
Pillow
sv-ttk
pyperclip
